Time: 16 ms (79.88%) | Memory: 44.8 MB (29.47%) - LeetSync

+++ 2227-sum-of-subarray-ranges/sum-of-subarray-ranges.java
@@ -3,8 +3,8 @@ class Solution {
         int n = nums.length;
         long res = 0;
         for(int i=0; i<n; i++){
-            long minV = nums[i];
-            long maxV = nums[i];
+            int minV = nums[i];
+            int maxV = nums[i];
             for(int j=i; j<n; j++){
                 minV = Math.min(minV, nums[j]);
                 maxV = Math.max(maxV, nums[j]);
